+++ design-system/docs/app/not-found.tsx
@@ -1,9 +1,15 @@
 import { Heading, Text } from '@voussoir/typography';
+import { Metadata } from 'next';
 import { DocsContent } from '../components/content';
+import { makePageTitle } from './utils';
+
+export const metadata: Metadata = {
+  title: makePageTitle('Page not found'),
+};
 
 export default function NotFoundPage() {
   return (
-    <DocsContent pageTitle="Page not found">
+    <DocsContent>
       <Heading size="large" elementType="h1">
         Page not found
       </Heading>
