+++ design-system/docs/app/utils.ts
@@ -1,0 +1,5 @@
+export const basePageTitle = 'Voussoir Design System';
+
+export function makePageTitle(title: string): string {
+  return `${title} | ${basePageTitle}`;
+}
