+++ design-system/docs/components/content/toc-context.tsx
@@ -1,3 +1,4 @@
+'use client';
 import {
   ComponentPropsWithoutRef,
   ProviderProps,
