+++ design-system/docs/utils/cache.ts
@@ -1,0 +1,4 @@
+import { cache as _cache } from 'react';
+
+export const cache: <T extends (...args: any[]) => any>(fn: T) => T =
+  typeof _cache === 'function' ? _cache : a => a;
