+++ design-system/packages/icon/src/index.ts
@@ -1,3 +1,4 @@
+'use client';
 export { Icon } from './Icon';
 
 export {
