+++ design-system/packages/layout/src/index.ts
@@ -1,3 +1,4 @@
+'use client';
 export { AspectRatio } from './AspectRatio';
 export { Box } from './Box';
 export { Divider } from './Divider';
