+++ design-system/packages/link/src/index.ts
@@ -1,3 +1,4 @@
+'use client';
 export {
   DefaultLinkComponent,
   LinkComponentContext,
