+++ design-system/packages/style/src/index.ts
@@ -1,3 +1,4 @@
+'use client';
 export { css, keyframes, injectGlobal } from '@emotion/css'; // simplify dependencies + ensure the same version of emotion is used
 
 export { transition } from './animation';
