+++ design-system/packages/typography/src/index.ts
@@ -1,3 +1,4 @@
+'use client';
 export { Heading, useHeadingContext } from './heading';
 export { Text, textClassList, useTextStyles, useTextContext } from './text';
